Add TrafficDividers component + minor code clean-up
