+++ envision/web/src/components/simulation.js
@@ -40,6 +40,7 @@ import Vehicles from "./vehicles.js";
 import DrivenPaths from "./driven_paths.js";
 import MissionRoutes from "./mission_routes.js";
 import Waypoints from "./waypoints.js";
+import TrafficDividers from "./traffic_dividers.js";
 
 import AgentScores from "./agent_scores";
 import earcut from "earcut";
@@ -68,8 +69,6 @@ export default function Simulation({
   const [roadNetworkBbox, setRoadNetworkBbox] = useState([]);
   const [laneDividerPos, setLaneDividerPos] = useState([]);
   const [edgeDividerPos, setEdgeDividerPos] = useState([]);
-  const [laneDividerGeometry, setLaneDividerGeometry] = useState([]);
-  const [edgeDividerGeometry, setEdgeDividerGeometry] = useState(null);
 
   const [worldState, setWorldState] = useState({
     traffic: [],
@@ -202,62 +201,12 @@ export default function Simulation({
       }
 
       setMapMeshes(meshes);
+
       GLTFLoader.UnregisterExtension(
         `load_gltf_extras_${worldState.scenario_id}`
       );
     });
   }, [scene, worldState.scenario_id]);
-
-  // Lane dividers
-  useEffect(() => {
-    if (scene == null || worldState.scenario_id == null) {
-      return;
-    }
-
-    for (const geom of laneDividerGeometry) {
-      geom.dispose();
-    }
-
-    let newLaneDividers = laneDividerPos.map((lines, idx) => {
-      let points = lines.map((p) => new Vector3(p[0], 0.1, p[1]));
-      let dashLine = MeshBuilder.CreateDashedLines(
-        `lane-divider-${idx}`,
-        { points: points, updatable: false, dashSize: 1, gapSize: 2 },
-        scene
-      );
-      dashLine.color = new Color4(...worldState.scene_colors["lane_divider"]);
-      return dashLine;
-    });
-
-    setLaneDividerGeometry(newLaneDividers);
-  }, [scene, JSON.stringify(laneDividerPos)]);
-
-  // Edge dividers
-  useEffect(() => {
-    if (scene == null || worldState.scenario_id == null) {
-      return;
-    }
-
-    if (edgeDividerGeometry != null) {
-      edgeDividerGeometry.dispose();
-    }
-
-    let edgeDividerPoints = edgeDividerPos.map((lines) => {
-      let points = lines.map((p) => new Vector3(p[0], 0.1, p[1]));
-      return points;
-    });
-
-    let newEdgeDividers = MeshBuilder.CreateLineSystem(
-      "edge-dividers",
-      { lines: edgeDividerPoints, updatable: false },
-      scene
-    );
-    newEdgeDividers.color = new Color4(
-      ...worldState.scene_colors["edge_divider"]
-    );
-
-    setEdgeDividerGeometry(newEdgeDividers);
-  }, [scene, JSON.stringify(edgeDividerPos)]);
 
   return (
     <div
@@ -277,7 +226,6 @@ export default function Simulation({
           height: "100%",
         }}
       />
-      <Bubbles scene={scene} worldState={worldState} />
       <Camera
         scene={scene}
         roadNetworkBbox={roadNetworkBbox}
@@ -289,6 +237,7 @@ export default function Simulation({
         vehicleRootUrl={`${client.endpoint.origin}/assets/models/`}
         egoView={egoView}
       />
+      <Bubbles scene={scene} worldState={worldState} />
       <DrivenPaths
         scene={scene}
         worldState={worldState}
@@ -301,6 +250,12 @@ export default function Simulation({
         worldState={worldState}
         egoWaypointModel={egoWaypointModel}
         socialWaypointModel={socialWaypointModel}
+      />
+      <TrafficDividers
+        scene={scene}
+        worldState={worldState}
+        laneDividerPos={laneDividerPos}
+        edgeDividerPos={edgeDividerPos}
       />
       {showScores ? (
         <AgentScores
